feat: Align all entities with DDL schema - fix 19 mismatches
Critical Type Fixes:
- ActivityResource.MediaFileId: Guid -> int (FK to media_files.file_id)
- StudentBadge.StudentBadgeId: int -> long (BIGINT for high volume)
- StudentProgress.ProgressId: int -> long (BIGINT for high volume)
- StudentBadge.EarnedAt: DateOnly -> DateTimeOffset (TIMESTAMPTZ)

Non-Nullable FK Fixes (NOT NULL in DDL):
- Module.ModuleTypeId: short? -> short
- Module.DifficultyLevel: int? -> int
- Activity.ActivityTypeId: short? -> short
- Badge.BadgeTypeId: short? -> short
- Badge.RarityId: short? -> short
- Badge.PointsValue: int? -> int (DEFAULT 0)
- Badge.DisplayOrder: int? -> int (DEFAULT 0)
- Badge.UnlockCriteria: string? -> string (JSONB NOT NULL)

Missing Properties Added:
- ActivityType.IsActive
- ResourceType.IsActive
- UsageContext.IsActive
- BadgeType.IsActive
- BadgeRarity.IsActive
- ProgressStatus.TenantId (NOT NULL FK)
- StudentBadge.ProgressId, CreatedBy, UpdatedBy

Configuration & Handler Updates:
- Updated all handlers for non-nullable FKs with default values
- Fixed ProgressStatus.StudentProgresses navigation property
- Added Tenant FK relationship to ProgressStatus
- Updated StudentBadge DTOs to use DateTimeOffset
- Updated ActivityResource handlers for int MediaFileId

Build: 0 errors, 4 warnings (null reference in ActivityResource handlers)

+++ src/Escola.Domain/Game/StudentProgress.cs
@@ -8,7 +8,7 @@ public class StudentProgress
     /// <summary>
     /// Internal database ID.
     /// </summary>
-    public int ProgressId { get; set; }
+    public long ProgressId { get; set; }
 
     /// <summary>
     /// External UUID for API exposure.
